requirements/base.txt
pytz==2024.1  # https://github.com/stub42/pytz
python-slugify==6.1.1  # https://github.com/un33k/python-slugify
Pillow==10.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.6.0  # https://github.com/evansd/whitenoise

# Django
# ------------------------------------------------------------------------------
django==4.2.11  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.4.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.61.1  # https://github.com/pennersr/django-allauth
django-crispy-forms==2.1  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==2024.2  # https://github.com/django-crispy-forms/crispy-bootstrap5

django-tables2==2.7.0
django-treebeard==4.7.1
sorl-thumbnail==12.10.0
django_bootstrap5==23.4
django_filter==23.5
django-tenants==3.6.1
qrcode==7.4.2
django-taggit==4.0.0
django-persistent-filters==1.0
django-htmx==1.17.3
wagtail==6.1.*
tablib==3.5.0
wagtailmenus==4.0
babel==2.9.1 # newer version fails because of langauge variants https://github.com/python-babel/babel/issues/454
django-money==3.4.1
django-autocomplete-light==3.11.0
mjml-python==1.3.3
mrml==0.1.14
djangoql==0.18.1
